+++ samplesheets/tests/test_plugins_taskflow.py
@@ -1,25 +1,31 @@
 """Tests for plugins in the samplesheets app with Taskflow enabled"""
 
+import os
+
 from irods.ticket import Ticket
 
+from django.forms.models import model_to_dict
 from django.test import RequestFactory, override_settings
 
 # Projectroles dependency
 from projectroles.app_settings import AppSettingAPI
 from projectroles.models import SODAR_CONSTANTS
-from projectroles.plugins import ProjectAppPluginPoint
+from projectroles.plugins import ProjectAppPluginPoint, PluginAPI
 
 # Taskflowbackend dependency
 from taskflowbackend.tests.base import TaskflowViewTestBase, IRODS_GROUP_PUBLIC
 
+from samplesheets.plugins import IRODS_STATS_CACHE_NAME, EMPTY_IRODS_STATS
 from samplesheets.tests.test_io import SampleSheetIOMixin, SHEET_DIR
 from samplesheets.tests.test_views_taskflow import (
     SampleSheetPublicAccessMixin,
     SampleSheetTaskflowMixin,
 )
+from samplesheets.views import MISC_FILES_COLL
 
 
 app_settings = AppSettingAPI()
+plugin_api = PluginAPI()
 
 
 # SODAR constants
@@ -32,6 +38,7 @@ APP_SETTING_SCOPE_PROJECT = SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT']
 # Local constants
 APP_NAME = 'samplesheets'
 SHEET_PATH = SHEET_DIR + 'i_small.zip'
+TEST_OBJ_NAME = 'test1.txt'
 
 
 class SamplesheetsModifyAPITestMixin:
@@ -61,20 +68,24 @@ class SamplesheetsModifyAPITestMixin:
             self.assertIsNone(ticket)
 
 
-class TestPerformProjectModify(
+class SamplesheetsPluginTaskflowTestBase(
     SamplesheetsModifyAPITestMixin,
     SampleSheetIOMixin,
     SampleSheetPublicAccessMixin,
     SampleSheetTaskflowMixin,
     TaskflowViewTestBase,
 ):
-    """Tests for perform_project_modify()"""
+    """Base class for samplesheets plugin tests with taskflow and iRODS"""
+
+    def _set_up_investigation(self):
+        """Set up investigation with iRODS collections"""
+        self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
+        self.study = self.investigation.studies.first()
+        self.assay = self.study.assays.first()
+        self.make_irods_colls(self.investigation)
 
     def setUp(self):
         super().setUp()
-        self.plugin = ProjectAppPluginPoint.get_plugin('samplesheets')
-
-        # Make project with owner in Taskflow without public guest access
         self.project, self.owner_as = self.make_project_taskflow(
             title='TestProject',
             type=PROJECT_TYPE_PROJECT,
@@ -82,13 +93,15 @@ class TestPerformProjectModify(
             owner=self.user,
         )
         self.sample_path = self.irods_backend.get_sample_path(self.project)
-        # Import investigation
-        self.investigation = self.import_isa_from_file(SHEET_PATH, self.project)
-        self.study = self.investigation.studies.first()
-        self.assay = self.study.assays.first()
-        # Create iRODS collections
-        self.make_irods_colls(self.investigation)
 
+
+class TestPerformProjectModify(SamplesheetsPluginTaskflowTestBase):
+    """Tests for ProjectAppPlugin.perform_project_modify()"""
+
+    def setUp(self):
+        super().setUp()
+        self._set_up_investigation()
+        self.plugin = ProjectAppPluginPoint.get_plugin('samplesheets')
         # Create dummy request
         self.req_factory = RequestFactory()
         self.request = self.req_factory.get('/')
@@ -317,25 +330,12 @@ class TestPerformProjectModify(
         self.assert_ticket_access(self.project, False)
 
 
-class TestPerformProjectSync(
-    SamplesheetsModifyAPITestMixin,
-    SampleSheetIOMixin,
-    SampleSheetPublicAccessMixin,
-    SampleSheetTaskflowMixin,
-    TaskflowViewTestBase,
-):
-    """Tests for perform_project_sync()"""
+class TestPerformProjectSync(SamplesheetsPluginTaskflowTestBase):
+    """Tests for ProjectAppPlugin.perform_project_sync()"""
 
     def setUp(self):
         super().setUp()
         self.plugin = ProjectAppPluginPoint.get_plugin('samplesheets')
-        self.project, self.owner_as = self.make_project_taskflow(
-            title='TestProject',
-            type=PROJECT_TYPE_PROJECT,
-            parent=self.category,
-            owner=self.user,
-        )
-        self.sample_path = self.irods_backend.get_sample_path(self.project)
 
     def test_sync(self):
         """Test perform_project_sync()"""
@@ -441,3 +441,107 @@ class TestPerformProjectSync(
             '',
         )
         self.assert_ticket_access(self.project, False, ticket_str)
+
+
+class TestUpdateIrodsStatsCache(SamplesheetsPluginTaskflowTestBase):
+    """Tests for ProjectAppPlugin.update_irods_stats_cache()"""
+
+    def setUp(self):
+        super().setUp()
+        self._set_up_investigation()
+        self.plugin = ProjectAppPluginPoint.get_plugin('samplesheets')
+        self.cache_backend = plugin_api.get_backend_api('sodar_cache')
+        self.cache_args = [
+            self.project,
+            self.irods_backend,
+            self.cache_backend,
+            self.irods,
+            self.user,
+        ]
+
+    def test_update_irods_stats_cache(self):
+        """Test update_irods_stats_cache() with no file in project"""
+        res = self.plugin.update_irods_stats_cache(*self.cache_args)
+        expected = {
+            'id': res.pk,
+            'project': self.project.pk,
+            'app_name': APP_NAME,
+            'name': IRODS_STATS_CACHE_NAME.format(uuid=self.project.sodar_uuid),
+            'user': self.user.pk,
+            'sodar_uuid': res.sodar_uuid,
+            'data': EMPTY_IRODS_STATS,
+        }
+        self.assertEqual(model_to_dict(res), expected)
+
+    def test_update_irods_stats_cache_file(self):
+        """Test update_irods_stats_cache() with file in project"""
+        misc_coll = self.irods.collections.create(
+            os.path.join(self.sample_path, MISC_FILES_COLL)
+        )
+        irods_obj = self.make_irods_object(misc_coll, TEST_OBJ_NAME)
+        self.make_checksum_object(irods_obj)
+        res = self.plugin.update_irods_stats_cache(*self.cache_args)
+        self.assertEqual(res.data, {'file_count': 1, 'total_size': 1024})
+
+    def test_update_irods_stats_no_coll(self):
+        """Test update_irods_stats_cache() with no sample collection"""
+        self.irods.collections.remove(self.sample_path)
+        res = self.plugin.update_irods_stats_cache(*self.cache_args)
+        self.assertEqual(res.data, EMPTY_IRODS_STATS)
+
+    def test_update_irods_stats_cache_existing(self):
+        """Test update_irods_stats_cache() with existing item"""
+        self.cache_backend.set_cache_item(
+            app_name=APP_NAME,
+            name=IRODS_STATS_CACHE_NAME.format(uuid=self.project.sodar_uuid),
+            data=EMPTY_IRODS_STATS,
+            project=self.project,
+            user=self.user,
+        )
+        misc_coll = self.irods.collections.create(
+            os.path.join(self.sample_path, MISC_FILES_COLL)
+        )
+        irods_obj = self.make_irods_object(misc_coll, TEST_OBJ_NAME)
+        self.make_checksum_object(irods_obj)
+        res = self.plugin.update_irods_stats_cache(*self.cache_args)
+        self.assertEqual(res.data, {'file_count': 1, 'total_size': 1024})
+
+
+# TODO: Add tests for ProjectAppPlugin.update_assay_shortcut_cache()
+# TODO: Add tests for ProjectAppPlugin.update_cache()
+
+
+class TestGetCategoryStats(SamplesheetsPluginTaskflowTestBase):
+    """Tests for ProjectAppPlugin.get_category_stats()"""
+
+    def setUp(self):
+        super().setUp()
+        self._set_up_investigation()
+        self.plugin = ProjectAppPluginPoint.get_plugin('samplesheets')
+
+    # NOTE: For sample count tests, see test_plugins
+
+    def test_get_category_stats(self):
+        """Test get_category_stats()"""
+        res = self.plugin.get_category_stats(self.category)
+        self.assertEqual(len(res), 3)
+        self.assertEqual(res[0].title, 'Samples')
+        self.assertEqual(res[1].title, 'Files')
+        self.assertEqual(res[1].value, 0)
+        self.assertEqual(res[1].unit, None)
+        self.assertEqual(res[2].title, 'Data')
+        self.assertEqual(res[2].value, 0)
+        self.assertEqual(res[2].unit, 'bytes')
+
+    def test_get_category_stats_file(self):
+        """Test get_category_stats() with file in project"""
+        misc_coll = self.irods.collections.create(
+            os.path.join(self.sample_path, MISC_FILES_COLL)
+        )
+        irods_obj = self.make_irods_object(misc_coll, TEST_OBJ_NAME)
+        self.make_checksum_object(irods_obj)
+        res = self.plugin.get_category_stats(self.category)
+        self.assertEqual(res[1].value, 1)
+        self.assertEqual(res[1].unit, None)
+        self.assertEqual(res[2].value, 1)
+        self.assertEqual(res[2].unit, 'KB')